+++ databento/common/dbnstore.py
@@ -36,7 +36,6 @@ from databento.common.data import SCHEMA_STRUCT_MAP
 from databento.common.error import BentoError
 from databento.common.iterator import chunk
 from databento.common.symbology import InstrumentMap
-from databento.common.symbology import SymbolInterval
 from databento.common.validation import validate_enum
 from databento.common.validation import validate_file_write_path
 from databento.common.validation import validate_maybe_enum
@@ -812,23 +811,13 @@ class DBNStore:
                 raise ValueError("a schema must be specified for mixed DBN data")
             schema = self.schema
 
-        record_type = SCHEMA_STRUCT_MAP[schema]
-        record_iter = filter(lambda r: isinstance(r, record_type), self)
-
-        if map_symbols:
-            self._instrument_map.insert_metadata(self.metadata)
-            symbol_map = self._instrument_map._data
-        else:
-            symbol_map = None
-
         with open(path, "xb") as output:
             self._transcode(
                 output=output,
-                records_iter=record_iter,
                 encoding=Encoding.CSV,
                 pretty_px=pretty_px,
                 pretty_ts=pretty_ts,
-                symbol_map=symbol_map,
+                map_symbols=map_symbols,
                 compression=compression,
                 schema=schema,
             )
@@ -1025,23 +1014,13 @@ class DBNStore:
                 raise ValueError("a schema must be specified for mixed DBN data")
             schema = self.schema
 
-        record_type = SCHEMA_STRUCT_MAP[schema]
-        record_iter = filter(lambda r: isinstance(r, record_type), self)
-
-        if map_symbols:
-            self._instrument_map.insert_metadata(self.metadata)
-            symbol_map = self._instrument_map._data
-        else:
-            symbol_map = None
-
         with open(path, "xb") as output:
             self._transcode(
                 output=output,
-                records_iter=record_iter,
                 encoding=Encoding.JSON,
                 pretty_px=pretty_px,
                 pretty_ts=pretty_ts,
-                symbol_map=symbol_map,
+                map_symbols=map_symbols,
                 compression=compression,
                 schema=schema,
             )
@@ -1114,27 +1093,33 @@ class DBNStore:
     def _transcode(
         self,
         output: BinaryIO,
-        records_iter: Iterator[DBNRecord],
         encoding: Encoding,
         pretty_px: bool,
         pretty_ts: bool,
-        symbol_map: dict[int, list[SymbolInterval]] | None,
+        map_symbols: bool,
         compression: Compression,
         schema: Schema,
     ) -> None:
+        if map_symbols:
+            self._instrument_map.insert_metadata(self.metadata)
+            symbol_map = self._instrument_map._data
+        else:
+            symbol_map = None
+
         transcoder = Transcoder(
             file=output,
             encoding=encoding,
             compression=compression,
             pretty_px=pretty_px,
             pretty_ts=pretty_ts,
+            map_symbols=map_symbols,
             has_metadata=True,
             symbol_map=symbol_map,  # type: ignore [arg-type]
             schema=schema,
         )
 
         transcoder.write(bytes(self.metadata))
-        for records in chunk(records_iter, 2**16):
+        for records in chunk(self, 2**16):
             for record in records:
                 transcoder.write(bytes(record))
             transcoder.flush()