+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.22.1 - 2023-10-24
+
+#### Bug fixes
+- Fixed an issue where `DBNStore.to_csv` and `DBNStore.to_json` were mapping symbols even when `map_symbols` was set to `False`
+- Fixed an issue where empty symbology mappings caused a `ValueError` when loading symbols into the `DBNStore` instrument map
+
 ## 0.22.0 - 2023-10-23
 
 #### Enhancements